refactor(stella-store): move the global spend report into its own module
`usage.rs` is a grandfathered god file sitting close to its baseline
ceiling, and #4171's `floor_executions` — a field, its doc, and the
paragraph explaining why the count reads the rollup join and not
`telemetry.usage_complete` — pushed it 29 lines over. The ceiling is not
the thing to move: `make file-size-update` would turn the gate green while
making the file worse.

`global_telemetry_totals` and `GlobalTelemetryRow` move to
`usage/global_report.rs`, beside `usage/tool_report.rs`, which is the same
shape for the same reason — one cross-project reader, its row type, and
the reasoning behind its query.

No witness: a move, and the existing tests exercise it unchanged.

Refs #4171
